Applied patch from Ryan which introduces the check_ssh script and added it to the Makefile.
"Here is my first attempt at a check_ssh module.  As listed in the TODO,
I would like to find a cleaner implimentation of dealing with specified
directive values and default directive values, but figured it was time
to submit this for review/discussion.

This is probably taylored towards OpenSSH (since that is what I
have access to), but I am hoping it is general enough for other SSH
daemons as well.  If not, I have no problem with the module being
renamed to check_openssh, etc.

This has been tested on RedHat, Debian, HP-UX 10.20, HP-UX 11.00, and
HP-UX 11.11. "